Revert "Updated ME67 and added additional unit tests  (#945)"
This reverts commit 5e77873abc1c451dbea8f6449d36966078c2bf39.
